Remove flowtables from the veth interface
Keeping flowtables for the veth interface causes routing failure for
consecutive packages. After the initial packet hits the prerouting chain
and perform the dnat, we want it to continue with the post routing and
then resubmitted immediately to the kernel for ESP encapsulation.
However, with flowtable, the packets are immediately transformed to
their pre ESP phase (having private addresses in source and destination)
and then submitted back to veth interface. This causes packets to leak
out to vetho* interface, therefore, we miss the chance of creating an
ESP packet. This caused failures for balancing, therefore, we limit the
flowtables to the tap* device.

This change increases the latency for both lb vms and non lb vms by 3%.
It impacts the whole fleet because nftables firewall table is used by
every type of VM. The test methodology that is used to find the 3%
latency increase is that;
1. Provision 2 postgres instances in the same host,
2. Install pgbench on the host.
3. Modify nftables for one of the postgres instances to not have veth*
interface for flowtables.
4. Run benchmarks like the following;
pgbench -i -s 100 PG_HOST_NAME
pgbench -c 1 -j 1 -T 600 -P 60 -S PG_HOST_NAME
5. Record the average latency from the output.
6. Repeat the step 4 for the other host.