Load Balancer DNS integration
We need to provide a hostname to the customers to be able to face a
single address to provide continuous connectivity even when the
underlying IP addresses of the load balancer changes. We accomplish this
by integrating load balancer with our existing DNS service. We will
create a new dns zone and manage individual host names depending on the
load balancer activity.