Add Load Balancer Health Probes
Health probes are an important part of a load balancer. The main
objective is to be able to find nodes that are down and remove them from
the load balancer so that the new connections are directed to the
healthy instances. In this commit, we implement health probes as a
separate Prog which is a child of the load balancer strand. This way,
the load balancer can update the state of individual nodes while the
health probes are individually performed. We create a single prog per
node in the load balancer. This way, a faulty node does not block the
health checks of other VMs. In case a status change is detected, we
start an update on the load balancer which reprograms all of the
nftables rules (including the failing one) to redirect the traffic to
only active VMs. In case the process is accessible again, we include it
in the load balancer.