VM destroy that is part of a load balancer is handled
When we need to destroy a VM, we need to be able to graciously remove it
from the load balancer, give enough time to the clients of the load
balancer to start using other nodes. For that, we are removing most of
the entities at the VM destroy time but still keep the network namespace
running for 10 more minutes. That is more than enough for dns to expire
and all of the connections to be rerouted to the other load balancer
nodes.