Remove name from pseudo-storage-size
We are trying to make the fields we are sending to web endpoints as close as to
the fields we are sending to api endpoints. We don't have pseudo-storage-size
fields in the api endpoints. By not setting the name field, we are preventing
it to be send as a parameter to backend.