Rename storage-size as storage_size
We have conflicting naming conventions in web and api endpoints (underscore vs.
dash as a separator). This already causes confusion in error messages generated
from validations. However, the conflict becomes more apparent when we try to
combine web and api endpoints. Since the API is public, we are using it as the
converged format and rename storage-size to storage_size