Combine business logic for Postgres resource creation endpoints
Most of the code is straightforward changes to merge the logic in two branches
in the postgres_helper. However there is one interesting bit where we set the
location at web/project/postgres.rb. This is because the API request for
resource creation already has location in the URL. However for web endpoints
the location is selected from the UI and cannot be dynamically put to the
form's action. due to csrf checks. So we need to set the location there.