Assemble Pages with correct Prog
Assume ProgA registers a deadline and then pushes another ProgB. If that
deadline expires while ProgB is executing, the Page is created with a
reference to ProgB instead of ProgA. As a result, the Page has
misleading information and even if ProgA reaches the target state,
the Page is never resolved. This commit fixes this issue by walking
back the links on the stack to identify the Prog that has registered
the deadline.