Limit storage size options for VMs
We now only allow a default storage size based on instance size and 2x of the
default storage size for VMs. This is to reduce the number of options and to
make it easier for users to select the right storage size. It seems high number
of options can create confusion and adds yet another thing to decide on.