Limit storage size options for PostgreSQL
We now only allow a default storage size based on instance size, 2x and 4x of
the default storage size for PostgreSQL. This is to reduce the number of
options and to make it easier for users to select the right storage size. It
seems high number of options can create confusion and adds yet another thing
to decide on.