Allow setting firewall rules in bulk for firewalls
Firewalls have helper methods to add and remove rules, but they increment the
update_firewall_rules semaphore for each rule added or removed. This is fine in
terms of correctness, because the operation is done in transaction and other
entities will saw the change atomically. However, it is slower, because we need
to go to database multiple times. It also adds unnecessary entries for each
incremented semaphore. Instead, I added a bulk firewall rule set method to
handle this case more efficiently.