Fix strand delay calculation to use old schedule time instead of updated one
I've identified some inconsistencies between the VM creation time and
the strand schedule time. There are instances where the initial strand
schedule time and VM creation time differ significantly, yet the
schedule delay is minimal. This led me to look deeper to our strand
running code further.

We started to reload the strand at
bbf3e142bda5b408cd94150a435efca6bd40204c when we take the lease. This
ensures we have the most recent version of the strand. However, we also
update the try count and schedule time while taking the lease. Thus,
upon taking the lease, we have a new schedule time for the strand in
this context. We then calculate the delay based on strand's schedule
time, so the new schedule time. This results in us observing a smaller
delay than expected, or even a negative delay. As the try count
increases, this negative delay also increases. This bug also provides an
explanation for the the respirate picking jobs from the future mystery.

Fixes #1089

I moved the delay property from 'strand_started' to 'lease_acquired'. We
run the strand for 30 seconds when we acquire the lease. Therefore,
delays become meaningful when we first start running the strand.

It seems the respirate may not be keeping up with the current load, but
this issue has been overlooked due to this bug.