Manage PostgreSQL firewalls from the parent resource
Each PostgresServer used to manage its own firewall and we used to attach each
firewall to the private subnet. However, we were not properly cleaning up the
firewalls when the PostgresServer was deleted, causing the old rules in the
firewall to be active for that particular subnet. We could fix that particular
bug by better managing firewall of each PostgresServer (e.g. properly cleaning
them up after the deletion of the PostgresServer). However, it is better to
centralize the management of the firewalls in the parent entity, that is
PostgresResource.