Add custom inspect representation for UBID
When we get a UBID in pry, the displayed value is not useful at all. Usually, I
prefer to view the UBID or UUID and use "to_s" or "to_uuid" to retrieve the
value. I've updated the inspect method to display the UBID and UUID values by
default. I believe this will be helpful for operators.

Before:

    [1] clover-development(main)> UBID.parse("vmy2mfzyj1yj2kv7rpevcp9afs")
    => #<UBID:0x0000000112e55610 @value=319892216702515363489663739989322781948>

After:

    [1] clover-development(main)> UBID.parse("vmy2mfzyj1yj2kv7rpevcp9afs")
    => #<UBID:Vm  @ubid="vmy2mfzyj1yj2kv7rpevcp9afs" @uuid="f0a8fffa-41f4-8774-9ecf-8b3b6cb254fc">