Delete blob storage object when cache entry destroyed
When the cache entry is destroyed, we need to delete the blob storage
object too. We clean up the object in the `after_destroy` hook. If the
cache object is already deleted, we don't need to do anything.

Also, the cache might not be committed yet. So if the multipart upload
is not completed, we abort the upload. If the upload is aborted already,
we don't need to do anything.