Suppress new spec warnings
rspec displays a warning when an expectation is set on a nil object. We
can refactor this expectation to prevent this warning.

    WARNING: An expectation of `:firewalls` was set on `nil`. To allow
    expectations on `nil` and suppress this message, set
    `RSpec::Mocks.configuration.allow_message_expectations_on_nil` to
    `true`. To disallow expectations on `nil`, set
    `RSpec::Mocks.configuration.allow_message_expectations_on_nil` to
    `false`. Called from
    spec/model/postgres/postgres_resource_spec.rb:61:in `block (2
    levels) in <top (required)>'.

aws-sdk-s3 has begun to print a warning about an unused variable. We
already have a suppression in place for this library, and I've added
another one to it.

    ruby/3.2.4/lib/ruby/gems/3.2.0/gems/aws-sdk-core-3.201.1/lib/aws-sdk-core/cbor/encoder.rb:131:
    warning: assigned but unused variable - base