Aggregate GitHub usage by installation not project
We decided to aggregate daily GitHub usage to reduce the number of
billing records during implementation. Currently, we aggregated them by
project. However, since a project can have multiple installations, it's
difficult to determine which installation is responsible for the usage.
While the billing record already has a project ID column, it would be
more efficient to use the installation ID to aggregate usage.