Refactor LB web endpoints to remove health check params
We have decided to not make health check related params, configurable.
Therefore, we are removing those from the UI and related endpoints.

This commit also includes a small update to the default values. We are
going to run health checks every 30 seconds, with 15 seconds timeout.
The up threshold is 3 and the down threshold is 2 by default.