Update Load Balancer hostname to use subnet.ubid
Load Balancer names are unique in a given project. Furthermore, there
must be a private subnet that a load balancer is created under. Instead
of generating unique hostnames per load balancer, we are reusing the
private subnet ubid. This will be useful in future when we introduce the
TLS integration and manage certificates for the customer.