Load Balancer API Updates for the Terraform Provider
This commit tweaks some of the endpoints to accommodate Terraform
Provider implementation in an easier way. It also contains an additional
PATCH endpoint which is a much better approach than implementing
individual resource types to handle /attach-vm /detach-vm endpoints.
Another change is using the vm ubids in an array form instead of
serializing multiple VMs in the response body for a load balancer. It is
still used in the WEB endpoints but not in the APIs.