Add helper methods for quota calculations
I was not entirely sure about whether to put default quotas as config file or
as a database table. I decided to go with a config file for now, because
- It is easier to change. Does not require a migration.
- It is easier to view. You can see all the quotas in one place.
- It seldom changes. Control plane does not need to edit default quotas as
part of its normal operation.

The drawback is that, we cannot join with it, so I had to write a helper
function to calculate effective_quota_value. Though, it is small enough that
I'm not sure if the join version would be any better.