Add quota check for Github runner provisionings
We are calling quota_available? method with 0 as requested_additional_usage.

`project.quota_available?("GithubRunnerCores", 0)`

This looks weird, but it is intentional. In existing Github quota calculations,
we compare the total allocated cores with the core limit and allow passing the
limit once. This is because we check quota and allocate VMs in different labels
(hence different transactions) and it is difficult to enforce the quotas ih the
environment with lots of concurrent requests. There are some remedies, but it
would require some refactoring that I'm not keen to do at the moment. Passing 0
as requested_additional_usage keeps the existing behavior.