Add different quotas for different reputation levels
Currently we have 2 levels of reputation: new and verified. Each have different
default quotas. I think we can have more levels of reputation such as sponsored
or enterprise. Depending on the reputation level, we can have different default
quotas for each of them.

I'm not entirely sold on the naming "reputation", "type" feels more appropriate
but it is too generic, so I didn't want to use it.