Add certain colors and width percentages to safelist
While rendering the progress bar, we set the width property of it to percentage
value. We cannot do it via inline style because of the security policy. We need
to set it via tailwind class. However, since the value is dynamic and tailwind
class is not known at the compile time, thus not added to the generated css. To
solve this, we add certain colors and width percentages to the safelist.