Combine create_unix_user commands to single command
We cache SSH connections to our host to run commands more quickly. Even
for short commands, the network adds ~0.4s of overhead. Previously, we
achieved a few seconds of gain by merging multiple commands into a
single one in GithubRunner, which makes a significant difference in
time-sensitive operations.

We can also consolidate the `create_unix_user` commands into a single
command. Additionally, I have relocated `sudo usermod -a -G kvm #{q_vm}`
from the `prep` label to this command, as it is more logical to execute
it here. I will remove it from the `prep` label in the next commit. If
there are any strands that have already passed `create_unix_user` but
haven't run `prep` yet. Since this operation is idempotent, it's safe to
run it multiple times.