Reduce sleep time in vm provisioning from 12 to 10 seconds
In the `wait_sshable` label, we initially set a long sleep time because
we expected the first run to take some time. The sleep time was
initially set at 12 seconds. However, we noticed that some duration
times were approaching this 12-second mark. As a result, we decided to
decrease the sleep time to 10 seconds.