Move the trigger for starting vm's systemd unit to rhizome
After the `prep` label is complete, the control plane run the systemd
unit in the `run` label. The `prep` script is daemonized, which causes
the control plane to poll too frequently to check its status, then it
starts the systemd unit. When we move this run command to the end of the
prep script at rhizome, the vm starts to boot immediately after it is
prepared. Thus, it will start booting even before the control plane
acknowledges that the prep is complete.

Our main goal is to minimize the interaction between the control plane
and the data plane to enhance our provisioning times. By removing the
`run` label and initiating the booting process earlier, I anticipate a
1-2 second improvement in provisioning time.

`VmSetup.recreate_unpersisted` is invoked in
`Vm::Nexus.start_after_host_reboot` and shares similar logic with
`VmSetup.prep`. Although it's not time-sensitive, I've changed it for
consistency.

I remove the `run` label in the next commit for smoother deployment.