Condense pagination parameters in openapi.json
Using a "component"[^1] it's possible to make reference to the three
parameters that are constituents of pagination in formation, in one
line.

It does raise further questions, if such parameters can/should
*always* be accepted in the API (or with few exceptions that are
isolated out) so that it can be repeated even less.  But this is a
better-coupled and more terse improvement nonetheless.

[^1]: https://swagger.io/docs/specification/components/