Do not invite deleted accounts to the project
Typically, user accounts are identified uniquely by their email.
However, this only applies to active accounts. When an account is
deleted, Rodauth changes its status to '3:deleted' and allows the
creation of a new account with the same email. Therefore, when we select
a user solely by email, we might retrieve a deleted account. To prevent
this, we need to exclude deleted accounts. Some of our customers have
already encountered this issue.