Destroy account when it's closed
Rodauth marks accounts as deleted when they are closed but not actually
delete records from the accounts table. It has
`delete_account_on_close?` option [^1] that can be set to `true` to
actually delete records from the accounts table when they are closed.
But it calls `.delete` method on account dataset [^2]. It has two
drawbacks:
  - It doesn't delete access tags of the account in the projects.
  - It calls `.delete` instead of `.destroy`. So the account doesn't
    archived to `deleted_records` table. We want to archive it for audit
    purposes.

So I override `delete_account` method to solve these drawbacks.

[^1]: https://github.com/jeremyevans/rodauth/blob/2412336a6107feb5e8e3f107c9dc50c781a5d42d/lib/rodauth/features/close_account.rb#L48
[^2]: https://github.com/jeremyevans/rodauth/blob/2412336a6107feb5e8e3f107c9dc50c781a5d42d/lib/rodauth/features/close_account.rb#L77