Set Content-Type of errors to application/json in API
`CloverBase` has some code to handle content conversion of errors (in
`clover_base.rb:parse_error`, but the `Content-Type` stays the default
`text/html`, which is wrong for the API.

Set the type to `application/json`.

I found this while verifying an openapi schema, which requires content
type matching to conform properly.