Decouple cloudinit and networking setup in vm prepare
I'm planning to run a few vm preparation steps in parallel. To achieve
this, we need to decouple these steps as much as possible. The
`cloudinit` uses `guest_ephemeral` from the `setup_networking` step. In
reality, this is a subdivide of the gua and we've calculated it in
various places. Therefore, we can also calculate it here.