Run networking and storage setup in parallel during vm prep
Setting up networking and storage are the most time-consuming steps in
our VM preparation script, taking up the majority of the process.
Ideally, each step takes 1 second, but when we have too many active vm
provisioning, each can take up to 5-8 seconds. However, as these steps
are independent, we can execute them in parallel to save time. This
approach could save us 1 second in the best-case scenario and 5-8
seconds in the worst-case scenario.