Load Balancer Health Probs Protocol Support
This commit adds the health check protocol support. This is needed once
we start providing certificates. If the customer implements an
application that runs HTTPS endpoints, we must use that. For that, we
already provide the certificate to them. It's also possible that they
don't want the health checks to be performed at the TCP level without
any specific endpoint. This commit supports that as well.