Separate VM Destroy logic from VM Detach for LBs
We used to manage VMs that are being destroyed or detached through the
same state "evacuating". That is causing issues since there are
different steps to take for both.
1. For VMs that are being destroyed, keep the load balancer definition
in place but remove self from the pool until DNS TTL is expired, and
then remove the VM <-> Load Balancer relationship and destroy the VM
entity.

2. For VMs that are being detached, we need to update DNS immediately,
remove the load balancer definition, update all of the nodes immediately
and remove the VM <-> load balancer relationship.