Suppress exception stack trace for a expected exception in the test output
Recently, I enhanced the robustness of our GitHub cache reserving
endpoint in commit 3129af12. To verify this feature, I created a test
case that triggers an exception. Our web routes print a stack trace for
unknown exceptions [^1], so this test case was designed to expect an
exception. While the case was successful, it also printed the stack
trace of the exception, which cluttered our test output. This commit
resolves the issue by suppressing the stack trace, changing the
exception to an expected exception.

[^1]: https://github.com/ubicloud/ubicloud/blob/9a0ad0305f8d7d807baadc1b64004d35a5eee50a/routes/clover_base.rb#L35