Fix Load Balancer for IPv6
The Load Balancer was not working when there were only IPv6 only VMs in
the load balancing pool. The reason is, when there is only IPv6 only
VMs, the DNS resolves to an IPv6 address and the load balancer was
unable to route the traffic properly. Main reason of that bug is the
firewall rules implementation. When we started implementing firewalls,
we decided that the traffic in private network would be exempt from the
firewall rules. However, we made a concious decision to block packets
that blong to a NEW connection if they have the destination address of
the private ipv6. Because, it would be weird to have such a packet when
we are communicating in between different VMs. The whole private
networking logic depends on public IPs in pre forward hook + the IPSec
Tunnels.

Now, fast forward to the load balancer implementation, we DNAT the
packets destined to the public IPv6 of the VM and route them through the
Private IPv6 address. Therefore, when we evaluate the firewalls, just
before completing SNAT + IPSec Tunnelling, the firewalls see that there
is a NEW connection that is destined to the private IPv6 address of the
VM and drops the packet. With this commit, we change that logic to
accept such packets.