Add E2E tests for the vm pool
We maintain VM pools for the most commonly used runner labels to reduce
queue time for our customers. This code path is heavily used during
periods of high load. It's important to conduct E2E tests for this path.
Essentially, we create a new VM pool of size one, after downloading the
runner images. We initiate creation after the download, as the VM
designated for the pool requires the image. Once the VM is ready for the
pool, we adjust the VM pool size to zero. This approach assists us in
confirming that the VM is selected from the pool. At the end of the
test, the pool should contain zero VMs.