Add helper class for PrivateSubnet routes
The business logic in both web and api endpoints are combined in the helper
class. Currently the combination is made simply adding a branch for each option
and copying the business logic from each endpoint to their respective branches.
Rubocop also moved some common lines out of the branches, but other than that
the logic is exactly same. There is no change in any functionality. We did not
try to converge to single logic and eliminate the branches in this commit. This
is intentional as I wanted to keep this commit as boring and add a separate
commit with the more interesting convergence stuff.