Restart primary postgres server after failover
This is required for archive_mode change to take effect after failover. Better
solution would be setting archive_mode to on at the standby all the time and
changing the archival behavior using archive_command, which doesn't require
restarting the server. However, I'm not entirely sure about the logistics of
that especially right after the failover time yet, so I'm going with the
simpler solution for now.