Use more descriptive region names in UI
We have been using region names like "eu-central-h1" in the UI which is cryptic
for users. We decided to simplify them and use more descriptive names referring
to the actual location of the respective data centers. What we would do in case
of having multiple data centers in the location is an open question currently,
we would probably use the city name or a number suffix. However, it does not
look like we would have that problem soon.

We are keeping the names in the API as is, because we don't want to introduce a
breaking change at the moment.