PostgresTimeline picks the blob store depending on the region
We are introducing a second blob store specifically for the US region
and we should be able to selectively use that. Therefore, we are making
changes to the PostgresTimeline to pick the blob store that is created
for that region. For that, we have to pass the location to the
PostgresTimelineNexus.assemble because there is simply no other way to
reach the location information of that resource. Furthermore, we are
removing the Config elements that are set to access to the MinIO cluster
to create new key pairs per bucket. Instead, we start to read those
values directly from the MinIO entity.