Generate parameters for Nexus::assemble from allowed_optional_parameters
We used to directly use request_body_params with double splat operator to pass
parameters to Nexus::assemble. This is not a good practice because it opens the
door to pass any parameter to Nexus::assemble. With this change we start to
generate the parameters for Nexus::assemble from allowed_optional_parameters.