Change runner monitored memory to "available" rather than "free"
"Free" is defined very particularly in Linux systems, to mean space
not used by either page caches nor applications. As such, "free" tends
towards a small number of any long-lived system, as pages are kept in
the cache.

The "available" statistic subtracts out the page cache, it refers to
memory that can be made available to applications, e.g. it is more
related to whether the OOM killer will run. Most of the time, this is
the more useful metric.

Secondly, by running awk against the Linux procfs `/proc/meminfo`
file, the implementation can optimize out one process spawn, for
`free`, which itself reads `/proc/meminfo` anyway, as confirmed by
`strace`

Finally, I thought it was best to test the command that gets passed to
SSH.  In this trivial case, it doesn't add much, but if add an iota of
interpolation, I find rendering the concrete expected argument value
as useful in bug auditing.