Relax storage size checks during restore
As a result of compute scaling, the resource might end up in a non-standard
compute/storage configuration. Then, when user tries to restore from backup,
the operation fails because it fails the storage size check. With this change,
we start to allow non-standard compute/storage configurations during restore
only if it matches with the parent.

I think we should not allow non-standard configurations as a result of scaling
operations in the first place, but if we do, we should allow the same config
during restore as well.