Calculate earliest_restore_time based on the backups in blob storage
We used to cache the earliest_backup_completed_at in the database, with the
intention of updating it when we purge the backups. However, we decided to
purge backups with blob storage policy. Hence, there is no opportunity in the
code to update the earliest_backup_completed_at. This causes displaying wrong
potential restore times in the UI. To fix this, although it is slower, I'm
calculating the earliest_restore_time based on the backups in the blob storage.