Update reserve cache to return error if cache already exists
If the cache exists for the passed key and version, github itself
returns "cache already exists" with a message including scope, key
and version information. To mimic the behavior, returning and error
including same information.

Note that, such a scenario might occur if explicit save action used
on the workflow and that workflow is run multiple times.