Add the implicit_subquery extension
Sometime around 787be9b1cee393cd152e7135b8b035af98a5daf4, we ran into
an interesting behavior in Sequel, where raw-SQL datasets ignore
methods chained onto them.  Cosider the following two expressions:

    DB["select * from pg_class"].where { oid > 7 }.sql
    => "select * from pg_class"

    DB[:pg_class].where { oid > 7 }.sql
    => "SELECT * FROM "pg_class" WHERE ("oid" > $1::int4); [7]"

At the time, we worked around this with a small tweak to the code.
But, it always bothered me that the ignoring of `where` clauses could
have a disastrous effect in a multi-tenant application.

Some months later, I wrote in to express my concern about this:
https://github.com/jeremyevans/sequel/discussions/2208.  I basically
became convinced to enable the `implicit_subquery` plugin anyway.

The change to remove `.get` was not expected.  I'm still hashing it
out in the discussion, but it seems to be a design conflict with
`.get`'s support for expression generation.  `single_value` and
`first` avoids extra wrapping.