Update health probes to use the hostname and check return type
We are updating the health probes to use the hostname to perform the
health checks. This is important for HTTPS health checks because it is
possible the customers will only accept that virtual hosting endpoint,
instead of raw IP.

Since we are in the network namespace and do not need to perform
a lookup, we resolve the hostname automatically by using the --resolve
option of curl.

We also make a small fix here to transform response_code to an integer
and perform the check. This is necessary in case the response_code has
spaces or \n characters at the start or end of the response code.