Fix VM removal when its attached to a load balancer
We need to remove the vm user only after the cert server is removed
because it is used by cert server and deluser command fails otherwise.

The other case is that, we might start deleting VM and then the load
balancer is deleted before VM deletion completes. In that case, we
should perform a check at the last state to verify the load balancer is
still intact before removing VM from the load balancer.