Allow to pin the runners of a project to specific location
We use all locations for GitHub runners to increase our overall
utilization. However, some customers experience job failures in certain
locations and need to be pin them to a specific one. While we don't have
a blanket policy to assign all customers to a specific location, we have
decided to accommodate this if a customer has a specific requirement.