Add option to bypass runner deregistration and destroy vm immediately
When we attempt to destroy the runner, we also deregister it from
GitHub.  We wait to receive a 'not found' response for the runner. If
the runner is still running a job and, due to stale data, it gets
mistakenly hopped to destroy, this prevents the underlying VM from being
destroyed and the job from failing. However, in certain situations like
fraudulent activity, we might need to bypass this verification and
immediately remove the runner.

First I named the semaphore as `force_destroy`, but then I renamed it as
`skip_deregistration`.  The former name was misleading, because the
operator should increment the `destroy` semaphore too.