Fix Almalinux nftables usage
Commit 7303ec6240b8e62b5d2b25316ced5def4dff0057 made a change to how we
boot VMs by adding new commands to the cloud-init. However, we realized
that almalinux yaml files on cloud-init fail with ipv6 subnet. YAML
files are sensitive to ":". Fixing that with escaping surfaced another
issue. Almalinux doesn't come with nftables installed by default.
Therefore, we are installing it at the time of provisioning.