Make project invitation emails case insensitive
When a user registers using an invitation email with different casing,
the system fails to find the invitation because the email column is
case-sensitive. I reviewed the rodauth code and found that it uses a
`citext` column for emails to ensure case insensitivity [^1]. Given that
we match `accounts.email` with `project_invitation.email`, we also need
to make `project_invitations.email` case-insensitive.

[^1]: https://github.com/jeremyevans/rodauth/blob/24bc83ff54c66b0251e26facba63b8777f586eac/README.rdoc?plain=1#L317