Handle additional cases in GitHub installation callback
When a new user installs our GitHub app, GitHub redirects them to a
callback URL in our console to finalize the installation. We perform
some checks and create a GithubInstallation record if it doesn't already
exist. This redirection also occurs when a user updates the installation
via the GitHub UI. Recently, we identified a need to handle more
scenarios in the callback. The current implementation provides unclear
error messages, leading to user confusion. This PR introduces handling
for the following situations:

Typically, GitHub includes a code as a query parameter in the callback
URL. We use this code to obtain the OAuth token and verify the user's
installation. Although the code is mandatory as per GitHub's
documentation, we've encountered instances where it's absent for update
callbacks. Consequently, customers receive an installation failed error
message even when they already have a GithubInstallation record. To
address this, I've moved the code check to precede the creation of the
installation record. This way, we can provide clearer messages to users
if they encounter other issues. If they've already installed the app and
have access permission, we can redirect them to the project's runner
page.

Another issue we've recently faced is when a user, who isn't an
administrator of the GitHub organization, needs approval from the admin.
In such cases, GitHub sends a notification to the admin and redirects
the non-admin user to the callback URL with "setup_action=request" and
without an installation ID. When the admin approves the installation,
GitHub redirects the admin to the callback URL to finalize the
installation. This implies that the admin needs an Ubicloud account with
appropriate project permissions. This way, we can create a
GithubInstallation record for their project. I've added a more detailed
error message for this scenario and implemented a redirect to the
project's user management page. This allows the user to add the admin to
the project with ease. While this process isn't yet seamless, it's the
best solution we currently have.

Finally, I've refactored the fraud check slightly. Since we have the
current user, we can use it to check for suspension.